Fix CDC read error (#96)
CDCEvent reads the _id field into a string. By default, the Mongo client v2 will not convert the ObjectID to a string, so we get an error at read time. To fix this, we now set BSONOptions on the collection to allow this deserialization. This wasn't caught in the previous tests because the test code explicitly set the _id field to a string when creating test events, so the test events deserialized without error. I've also updated the TestWriteSuccess test to use a Mongo generated _id which reproduces the error.

Alternatively I considered changing CDCEvent.ID to a bson.ObjectID type, but since this struct is used in the REST client library I do not want to leak that Mongo type and dependency into the client. I also considered creating an internal cdcEvent struct identical to CDCEvent but with the bson.ObjectID type, then have the server deserialize into that struct and copy to the CDCEvent struct, but that seemed heavy and unnecessary.